% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emails.R
\name{emails}
\alias{emails}
\alias{get_emails}
\alias{print_emails}
\title{Obtain emails in the table}
\usage{
get_emails(tbl, sorted = TRUE, delimiter = ", ")

print_emails(tbl, sorted = TRUE, delimiter = ", ")
}
\arguments{
\item{tbl}{A tibble with one required column "email"}

\item{sorted}{Should the emails be sorted before returning?}

\item{delimiter}{Character(s) to use as a delimiter to when concatenating emails}
}
\description{
These functions collapse the emails in the provided \code{tbl} to
a comma-separated string that can be cut and paste into an email
client's "To:" field.
}
\section{Functions}{
\itemize{
\item \code{get_emails}: return an email list

\item \code{print_emails}: print the emails to the console
}}

\examples{
html_file <- system.file(package = "contactparser", "testdata", "example.html")
parsed <- parse_contact_table(html_file)

get_emails(parsed)

print_emails(parsed)
print_emails(parsed, delimiter = "\n")

}
\seealso{
parse_contact_table
}
