% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emails.R
\name{get_emails}
\alias{get_emails}
\title{Obtain emails in the table}
\usage{
get_emails(tbl, sorted = TRUE)
}
\arguments{
\item{tbl}{A tibble with one required column "email"}

\item{sorted}{Should the emails be sorted before returning?}
}
\description{
Obtain emails in the table
}
\examples{
{

html_file <- system.file(package = "contactparser", "testdata", "example.html")
parsed <- parse_contact_table(html_file)
get_emails(parsed)

}
}
\seealso{
parse_contact_table
}
